backend/requirements.txt
Django==4.2.*
djangorestframework
django-cors-headers
django-environ
PyJWT
requests
gunicorn
whitenoise
dj-database-url
psycopg2-binary
Pillow
